IT MIGHT WORK i just don't have enough RAM

+++ useCases/munich/src/main/java/de/tum/bgu/msm/data/geo/ZoneMuc.java
@@ -8,7 +8,7 @@ public class ZoneMuc extends ZoneImpl {
     private final AreaTypes.SGType areaType;
     private final double ptDistance_m;
 
-    public ZoneMuc(float id, float area, AreaTypes.SGType areaType, double initialPTDistance_m, Region region) {
+    public ZoneMuc(int id, float area, AreaTypes.SGType areaType, double initialPTDistance_m, Region region) {
 
         super(id, area, region);
         this.areaType = areaType;
